Find proper manage.py command
Modified to find the proper manage.py command by searching up the file tree for a file named "manage.py". User plugins are able to manually override this with their own path, as the script will now check if g:pony_manage_filename is set prior to setting the value.